requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=hd77b12b_0
aiohttp=3.9.0=py311h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
arrow-cpp=11.0.0=ha81ea56_2
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.1.0=py311haa95532_0
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
blas=1.0=mkl
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.5=py311h5bb9823_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotli-python=1.0.9=py311hd77b12b_7
bzip2=1.0.8=hcfcfb64_5
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2023.11.17=h56e8100_0
certifi=2023.11.17=py311haa95532_0
cffi=1.16.0=py311h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=py311h59b6b97_0
cryptography=41.0.7=py311h89fc84f_0
cuda-cccl=12.3.101=0
cuda-cudart=11.8.89=0
cuda-cudart-dev=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-libraries-dev=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvrtc-dev=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-profiler-api=12.3.101=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
datasets=2.12.0=py311haa95532_0
debugpy=1.8.0=py311h12c1d0e_1
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.6=py311haa95532_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
filelock=3.13.1=py311haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=py311h2bbff1b_0
fsspec=2023.10.0=py311haa95532_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h8cc25b3_3
glog=0.5.0=hd77b12b_0
gmpy2=2.1.2=py311h7f96b67_0
grpc-cpp=1.48.2=hfe90ff0_1
huggingface_hub=0.17.3=py311haa95532_0
icu=73.1=h6c2663c_0
idna=3.4=py311haa95532_0
imbalanced-learn=0.11.0=pyhd8ed1ab_0
importlib-metadata=7.0.0=pyha770c72_0
importlib_metadata=7.0.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.18.1=pyh7428d3b_3
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=py311haa95532_0
joblib=1.3.2=pyhd8ed1ab_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=5.5.0=py311h1ea47a8_0
kiwisolver=1.4.4=py311hd77b12b_0
krb5=1.20.1=h5b6d351_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcurand=10.3.4.101=0
libcurand-dev=10.3.4.101=0
libcurl=8.4.0=h86230a5_1
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.12=h56d1f94_1
libexpat=2.5.0=h63175ca_1
libffi=3.4.4=hd77b12b_0
libjpeg-turbo=2.0.0=h196d8e1_0
liblapack=3.9.0=5_hd5c7e75_netlib
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h8d14728_1
libsqlite=3.44.2=hcfcfb64_0
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libuv=1.44.2=h2bbff1b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
libzlib=1.2.13=hcfcfb64_5
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py311h2bbff1b_0
matplotlib=3.8.0=py311haa95532_0
matplotlib-base=3.8.0=py311hf62ec03_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpc=1.1.0=h7edee0f_1
mpfr=4.0.2=h62dcd97_1
mpir=3.0.0=hec2e145_1
mpmath=1.3.0=py311haa95532_0
msys2-conda-epoch=20160418=1
multidict=6.0.4=py311h2bbff1b_0
multiprocess=0.70.14=py311haa95532_0
munkres=1.1.4=py_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
networkx=3.1=py311haa95532_0
numexpr=2.8.7=py311h1fcbade_0
numpy=1.26.2=py311hdab7c0b_0
numpy-base=1.26.2=py311hd01c5d8_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.2.0=hcfcfb64_1
orc=1.7.4=h623e30f_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py311hf62ec03_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.0.1=py311h045eedc_0
pip=23.3.1=pyhd8ed1ab_0
platformdirs=4.1.0=pyhd8ed1ab_0
ply=3.11=py311haa95532_0
prompt-toolkit=3.0.41=pyha770c72_0
psutil=5.9.5=py311ha68e1ae_1
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=py311h8a3a540_1
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=23.2.0=py311haa95532_0
pyparsing=3.0.9=py311haa95532_0
pyqt=5.15.10=py311hd77b12b_0
pyqt5-sip=12.13.0=py311h2bbff1b_0
pysocks=1.7.1=py311haa95532_0
python=3.11.6=h2628c8c_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py311h2bbff1b_1
python_abi=3.11=4_cp311
pytorch=2.1.1=py3.11_cuda11.8_cudnn8_0
pytorch-cuda=11.8=h24eeafa_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py311haa95532_0
pywin32=306=py311h12c1d0e_2
pyyaml=6.0.1=py311h2bbff1b_0
pyzmq=25.1.2=py311h9250fbb_0
qt-main=5.15.2=h19c9488_10
re2=2022.04.01=hd77b12b_0
regex=2023.10.3=py311h2bbff1b_0
requests=2.31.0=py311haa95532_0
responses=0.13.3=pyhd3eb1b0_0
safetensors=0.4.0=py311hcbdf901_0
scikit-learn=1.3.2=py311h142b183_2
scipy=1.11.4=py311h0b4df5a_0
setuptools=68.2.2=pyhd8ed1ab_0
sip=6.7.12=py311hd77b12b_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h6c2663c_0
sqlite=3.44.2=hcfcfb64_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=py311haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.13=h5226925_1
tokenizers=0.13.3=py311h49fca51_0
torchaudio=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.3.3=py311ha68e1ae_1
tqdm=4.65.0=py311h746a85d_0
traitlets=5.14.0=pyhd8ed1ab_0
transformers=4.32.1=py311haa95532_0
typing-extensions=4.9.0=hd8ed1ab_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.18=py311haa95532_0
utf8proc=2.6.1=h2bbff1b_0
vc=14.3=h64f974e_17
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
wcwidth=0.2.12=pyhd8ed1ab_0
wheel=0.42.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=py311haa95532_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.5=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.9.3=py311h2bbff1b_0
zeromq=4.3.5=h63175ca_0
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_5
zstd=1.5.5=hd43e919_0
